requirements.txt
django==5.2.8
Pillow==12.0.0
numpy==1.26.4
onnxruntime==1.15.1
coloredlogs==15.0.1
flatbuffers==25.9.23
humanfriendly==10.0
packaging==25.0
protobuf==6.33.1
sympy==1.14.0
mpmath==1.3.0
pyreadline3==3.5.4
sqlparse==0.5.3
asgiref==3.11.0
tzdata==2025.2